feat: resolved merge conflict
